Add test for database reports
Adds utility functions to create groups and entries for them.

Add test to make sure the right amount of entries show up in the
database report for password health.

Add signal when password health calculation is done so the test harness
can wait on it; otherwise the test finishes too soon and doesn't get the
right result from the application.